+++ src/main/java/org/folio/entlinks/utils/HeaderUtils.java
@@ -1,20 +1,20 @@
 package org.folio.entlinks.utils;
 
 import java.util.Arrays;
+import java.util.Optional;
 import lombok.experimental.UtilityClass;
 import org.apache.kafka.common.header.Headers;
 
 @UtilityClass
 public class HeaderUtils {
 
-  public static String extractHeaderValue(String headerKey, Headers headers) {
+  public static Optional<String> extractHeaderValue(String headerKey, Headers headers) {
     if (headerKey == null || headers == null) {
-      return "";
+      return Optional.empty();
     }
     return Arrays.stream(headers.toArray())
       .filter(header -> headerKey.equalsIgnoreCase(header.key()) && header.value() != null)
       .findFirst()
-      .map(header -> new String(header.value()))
-      .orElse("");
+      .map(header -> new String(header.value()));
   }
 }